clean up a lot of the SUM code
