+++ ott/loader/sum/README.md
@@ -1,0 +1,10 @@
+loader:sum
+===========
+
+purpose: download Shared Use Mobility (SUM) data.
+         The [./config/app.ini](../../../config/app.ini) file controls the list of data sources.
+
+run: bin/sum_load (optional -ini <name>.ini)
+
+     
+
