+++ ott/loader/sum/sum_cache.py
@@ -1,0 +1,25 @@
+import os
+import logging
+log = logging.getLogger(__file__)
+
+from ott.utils.cache_base import CacheBase
+
+from ott.loader.sum.sobi.sobi_cache import SobiCache
+
+class SumCache(CacheBase):
+    """ Does a 'smart' cache of a SUM data
+    """
+    def __init__(self, force_update=False):
+        ''' check osm cache
+        '''
+        super(SumCache, self).__init__(section='sum')
+        sobi = SobiCache()
+        sobi.check_feed(force_update)
+
+
+def main():
+    #import pdb; pdb.set_trace()
+    sum = SumCache(force_update=object_utils.is_force_update())
+
+if __name__ == '__main__':
+    main()
